cmd/repo: merge repo_*.go files to repo.go
