+++ cmd/repo.go
@@ -14,9 +14,15 @@
 package cmd
 
 import (
+	"fmt"
 	"os"
+	"time"
 
+	"github.com/pingcap-incubator/tiup/pkg/localdata"
 	"github.com/pingcap-incubator/tiup/pkg/meta"
+	"github.com/pingcap-incubator/tiup/pkg/repository"
+	"github.com/pingcap-incubator/tiup/pkg/utils"
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
 
@@ -64,4 +70,289 @@ of components or the repository itself.`,
 		newRepoGenkeyCmd(env),
 	)
 	return cmd
+}
+
+// the `repo add` sub command
+func newRepoAddCompCmd(env *meta.Environment) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "add <component-id> <platform> <version> <file>",
+		Short: "Add a file to a component",
+		Long:  `Add a file to a component, and set its metadata of platform ID and version.`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) < 4 {
+				return cmd.Help()
+			}
+
+			return addCompFile(repoPath, args[0], args[1], args[2], args[3])
+		},
+	}
+
+	return cmd
+}
+
+func addCompFile(repo, id, platform, version, file string) error {
+	// TODO
+	return nil
+}
+
+// the `repo component` sub command
+func newRepoCompCmd(env *meta.Environment) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "component <id> <description>",
+		Short: "Create a new component in the repository",
+		Long:  `Create a new component in the repository, and sign with the local owner key.`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) < 2 {
+				return cmd.Help()
+			}
+
+			return createComp(repoPath, args[0], args[1])
+		},
+	}
+
+	return cmd
+}
+
+func createComp(repo, id, name string) error {
+	// TODO
+	return nil
+}
+
+// the `repo del` sub command
+func newRepoDelCompCmd(env *meta.Environment) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "del <component> [version]",
+		Short: "Delete a component from the repository",
+		Long: `Delete a component from the repository. If version is not specified, all versions
+of the given component will be deleted.
+Manifests and files of a deleted component will be removed from the repository,
+clients can no longer fetch the component, but files already download by clients
+may still be available for them.`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			compVer := ""
+			switch len(args) {
+			case 2:
+				compVer = args[1]
+			default:
+				return cmd.Help()
+			}
+
+			return delComp(repoPath, args[0], compVer)
+		},
+	}
+
+	return cmd
+}
+
+func delComp(repo, id, version string) error {
+	// TODO
+	return nil
+}
+
+// the `repo genkey` sub command
+func newRepoGenkeyCmd(env *meta.Environment) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "genkey",
+		Short: "Generate a new key pair",
+		Long:  `Generate a new key pair that can be used to sign components.`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			pubKey, privKey, err := genKeyPair()
+			if err != nil {
+				return err
+			}
+			// TODO: save keys to files
+			fmt.Printf("Private key generated:\n%s\n", privKey)
+			fmt.Printf("Public key of the private key:\n%s\n", pubKey)
+			return nil
+		},
+	}
+
+	return cmd
+}
+
+func genKeyPair() ([]byte, []byte, error) {
+	// TODO
+	return []byte("public key"), []byte("private key"), nil
+}
+
+// the `repo init` sub command
+func newRepoInitCmd(env *meta.Environment) *cobra.Command {
+	var (
+		pubKey  string // public key of root
+		privKey string // private key of root
+	)
+	cmd := &cobra.Command{
+		Use:   "init [path]",
+		Short: "Initialise an empty repository",
+		Long: `Initialise an empty TiUP repository at given path. If path is not specified, the
+current working directory (".") will be used.`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) == 1 {
+				repoPath = args[0]
+			}
+
+			// create the target path if not exist
+			if utils.IsNotExist(repoPath) {
+				var err error
+				if err = os.Mkdir(repoPath, 0755); err != nil {
+					return err
+				}
+			}
+			// init requires an empty path to use
+			empty, err := utils.IsEmptyDir(repoPath)
+			if err != nil {
+				return err
+			}
+			if !empty {
+				return errors.Errorf("the target path '%s' is not an empty directory", repoPath)
+			}
+
+			return initRepo(repoPath)
+		},
+	}
+
+	cmd.Flags().StringVar(&pubKey, "pubkey", "", "Path to the public key file")
+	cmd.Flags().StringVar(&privKey, "privkey", "", "Path to the private key file")
+
+	return cmd
+}
+
+func initRepo(path string) error {
+	currTime := time.Now()
+	repoManifests := repository.NewManifests(path)
+	// TODO: set key store
+
+	// initial manifests
+	newManifests := make([]*repository.Manifest, 0)
+
+	// init the root manifest
+	root := &repository.Root{
+		SignedBase: repository.SignedBase{
+			Ty:          "root",
+			SpecVersion: "TODO",
+			Expires:     currTime.UTC().Add(time.Hour * 24 * 365).Format(time.RFC3339), // 1y
+			Version:     1,                                                             // initial repo starts with version 1
+		},
+		Roles: make(map[string]*repository.RoleMeta),
+	}
+	rootManifest := &repository.Manifest{
+		Signed: root,
+	}
+	root.Roles[root.Filename()] = root.GetRole()
+	newManifests = append(newManifests, rootManifest)
+
+	// init index
+	index := &repository.Index{
+		SignedBase: repository.SignedBase{
+			Ty:          "index",
+			SpecVersion: "TODO",
+			Expires:     currTime.UTC().Add(time.Hour * 24 * 365).Format(time.RFC3339), // 1y
+			Version:     1,
+		},
+		Owners:            make(map[string]repository.Owner),
+		Components:        make(map[string]repository.Component),
+		DefaultComponents: make([]string, 0),
+	}
+	// TODO add initial owner to index.Owners
+	root.Roles[index.Filename()] = index.GetRole()
+	newManifests = append(newManifests, &repository.Manifest{
+		Signed: index,
+	})
+
+	// init snapshot
+	snapshot := &repository.Snapshot{
+		SignedBase: repository.SignedBase{
+			Ty:          "snapshot",
+			SpecVersion: "TODO",
+			Expires:     currTime.UTC().Add(time.Hour * 24).Format(time.RFC3339), // 1d
+			Version:     0,                                                       // not versioned
+		},
+	}
+	snapshotManifest := &repository.Manifest{
+		Signed: snapshot.SetVersions(newManifests),
+	}
+	root.Roles[snapshot.Filename()] = snapshot.GetRole()
+	newManifests = append(newManifests, snapshotManifest)
+
+	// init timestamp
+	timestamp := &repository.Timestamp{
+		SignedBase: repository.SignedBase{
+			Ty:          "timestamp",
+			SpecVersion: "TODO",
+			Expires:     currTime.UTC().Add(time.Hour * 24).Format(time.RFC3339), // 1d
+			Version:     uint(currTime.Unix()),
+		},
+	}
+	timestamp, err := timestamp.SetSnapshot(snapshot)
+	if err != nil {
+		return err
+	}
+	timestampManifest := &repository.Manifest{
+		Signed: timestamp,
+	}
+	root.Roles[timestamp.Filename()] = timestamp.GetRole()
+	newManifests = append(newManifests, timestampManifest)
+
+	// write to files
+	for _, m := range newManifests {
+		if err := repoManifests.SaveManifest(m); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// the `repo owner` sub command
+func newRepoOwnerCmd(env *meta.Environment) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "owner <id> <name>",
+		Short: "Create a new owner for the repository",
+		Long: `Create a new owner role for the repository, the owner can then perform management
+actions on authorized resources.`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			if len(args) < 2 {
+				return cmd.Help()
+			}
+
+			return createOwner(repoPath, args[0], args[1])
+		},
+	}
+
+	return cmd
+}
+
+func createOwner(repo, id, name string) error {
+	// TODO
+	return nil
+}
+
+// the `repo yank` sub command
+func newRepoYankCompCmd(env *meta.Environment) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "yank <component> [version]",
+		Short: "Yank a component in the repository",
+		Long: `Yank a component in the repository. If version is not specified, all versions
+of the given component will be yanked.
+A yanked component is still in the repository, but not visible to client, and is
+no longer considered stable to use. A yanked component is expected to be removed
+from the repository in the future.`,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			compVer := ""
+			switch len(args) {
+			case 2:
+				compVer = args[1]
+			default:
+				return cmd.Help()
+			}
+
+			return yankComp(repoPath, args[0], compVer)
+		},
+	}
+
+	return cmd
+}
+
+func yankComp(repo, id, version string) error {
+	// TODO
+	return nil
 }
