--- cmd/repo_del.go
@@ -1,49 +1,0 @@
-// Copyright 2020 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cmd
-
-import (
-	"github.com/pingcap-incubator/tiup/pkg/meta"
-	"github.com/spf13/cobra"
-)
-
-func newRepoDelCompCmd(env *meta.Environment) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "del <component> [version]",
-		Short: "Delete a component from the repository",
-		Long: `Delete a component from the repository. If version is not specified, all versions
-of the given component will be deleted.
-Manifests and files of a deleted component will be removed from the repository,
-clients can no longer fetch the component, but files already download by clients
-may still be available for them.`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			compVer := ""
-			switch len(args) {
-			case 2:
-				compVer = args[1]
-			default:
-				return cmd.Help()
-			}
-
-			return delComp(repoPath, args[0], compVer)
-		},
-	}
-
-	return cmd
-}
-
-func delComp(repo, id, version string) error {
-	// TODO
-	return nil
-}
