--- cmd/repo_yank.go
@@ -1,49 +1,0 @@
-// Copyright 2020 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cmd
-
-import (
-	"github.com/pingcap-incubator/tiup/pkg/meta"
-	"github.com/spf13/cobra"
-)
-
-func newRepoYankCompCmd(env *meta.Environment) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "yank <component> [version]",
-		Short: "Yank a component in the repository",
-		Long: `Yank a component in the repository. If version is not specified, all versions
-of the given component will be yanked.
-A yanked component is still in the repository, but not visible to client, and is
-no longer considered stable to use. A yanked component is expected to be removed
-from the repository in the future.`,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			compVer := ""
-			switch len(args) {
-			case 2:
-				compVer = args[1]
-			default:
-				return cmd.Help()
-			}
-
-			return yankComp(repoPath, args[0], compVer)
-		},
-	}
-
-	return cmd
-}
-
-func yankComp(repo, id, version string) error {
-	// TODO
-	return nil
-}
